refactor(name resolution): moving name resolution pass to its own subfolder
