+++ include/Ark/Compiler/Welder.hpp
@@ -27,7 +27,7 @@
 #include <Ark/Compiler/Package/ImportSolver.hpp>
 #include <Ark/Compiler/Macros/Processor.hpp>
 #include <Ark/Compiler/AST/Optimizer.hpp>
-#include <Ark/Compiler/NameResolutionPass.hpp>
+#include <Ark/Compiler/NameResolution/NameResolutionPass.hpp>
 
 namespace Ark
 {
